⚡ Bolt: Optimize OpenCV Laplacian precision

+++ products/desktop/src/photo_selector_toolbox/core/sharpness.py
@@ -93,9 +93,9 @@ def get_image_data(filepath: Path) -> Optional[np.ndarray]:
 
 def _calculate_noise_from_gray(gray: np.ndarray) -> float:
     """Estimates noise from a pre-loaded grayscale array using MAD of the Laplacian."""
-    laplacian = cv2.Laplacian(gray, cv2.CV_64F)
+    laplacian = cv2.Laplacian(gray, cv2.CV_32F)
     mad = np.median(np.abs(laplacian - np.median(laplacian)))
-    return mad / 0.6745
+    return float(mad / 0.6745)
 
 
 def _calculate_sharpness_from_gray(gray: np.ndarray, grid_size: int = 1) -> float:
@@ -112,14 +112,14 @@ def _calculate_sharpness_from_gray(gray: np.ndarray, grid_size: int = 1) -> floa
         cropped = gray[h_start:h_end, w_start:w_end]
 
     if grid_size <= 1:
-        return cv2.Laplacian(cropped, cv2.CV_64F).var()
+        return float(cv2.Laplacian(cropped, cv2.CV_32F).var())
 
     ch, cw = cropped.shape
     block_h = ch // grid_size
     block_w = cw // grid_size
 
     if block_h < 10 or block_w < 10:
-        return cv2.Laplacian(cropped, cv2.CV_64F).var()
+        return float(cv2.Laplacian(cropped, cv2.CV_32F).var())
 
     max_score = 0.0
     for r in range(grid_size):
@@ -129,7 +129,7 @@ def _calculate_sharpness_from_gray(gray: np.ndarray, grid_size: int = 1) -> floa
             x0 = c * block_w
             x1 = x0 + block_w
             block = cropped[y0:y1, x0:x1]
-            score = cv2.Laplacian(block, cv2.CV_64F).var()
+            score = float(cv2.Laplacian(block, cv2.CV_32F).var())
             if score > max_score:
                 max_score = score
     return max_score
